Switch to result objects and process pools
Introduce new Result object as return value from ufmt_file, which
can be used to return file path, whether it changed, and what the
resulting diff was. This allows using ufmt_file from an executor
and aggregating the results in the parent process.

ufmt_paths now uses an executor when running ufmt_file on each path
to check. This allows concurrent formatting of multiple files, which
cuts the overall runtime on the botorch project from ~15s to ~4s on
an M1 Mac Mini. Running `black --diff` on the same directory runs in
~3.5s, while `usort check` takes ~14s, so this is about as fast as
we can get it without resorting to native extensions for everything.

Fixes #4