Fix #94: Make sure fake stdin paths are resolvable
The previous fake path used by ufmt_paths when formatting stdin
was not resolvable on Windows, breaking any pre- or post-processor
that needed to resolve paths during formatting. This changes the
default fake path to remove the unallowed characters, and adds a
simple test case with a pre-processor. This test fails with old
behavior, and passes with the new fake path name.