Github Actions support (#107)
Adds a new Github Action spec to check formatting

The action performs the following:

- checkout the project
- setup python (if python-version specified)
- install black/usort (if versions specified)
- install ufmt (latest or specific version)
- runs python -m ufmt check against given path

Also adds a new "integrations" section to the user guide, and also documents the pre-commit support there. The appropriate sections of the user guide are now linked to from the readme.

Documented examples depend on adding a new action-v1 tag to the repo (named so that it doesn't conflict with actual version tags in the changelog).