Start formatting files before walking is finished
Adds a simple generator function to yield paths to format,
and passes that to trailrunner so that it can start formatting
files while the generator is still walking and yield paths.
Manually pumps the first two items in order to retain the
base case of not using a child process if only a single file
needs to be formatted.

Fix #168