Pass target file path to ufmt config factory
This fixes `ufmt_file` to pass the full file path to the ufmt config
factory, rather than just the path's parent. This enables alternate
config factories to make decisions based on the target filename, suffix,
etc that can't be made just based on the parent path. This also aligns
with config factory usage for black and usort.

Fixes #196

[ghstack-poisoned]